app/requirements.txt
Flask==2.2.5
mysql-connector-python==8.1.0
pytest==7.4.0
flake8==6.1.0
